e2e: remove thin volume and thin pool
pvc-autoresizer uses TopoLVM's example manifests.  In this example,
a thin pool and a thin volume are also created.  They consume so much
storage capacity that the PV cannot be created in the pvc-autoresizer's
e2e test due to the lack of space.

The thin pool and the thin volume are not used in e2e test,
so I remove these resources to increase the available capacity.

Signed-off-by: Shinya Hayashi <shinya-hayashi@cybozu.co.jp>
Co-authored-by: Yuma Ogami <yuma-ogami@cybozu.co.jp>