docs: remove "Project Status" from README.md
pvc-autoresizer is already used in the production environments, the
project status in README.md does not reflect the actual status of this
project. The current maintainers have agreed to remove this sentence
from README.md. This is because the project status decided by the
project owners themselves is not very useful for the current stage and
might confuse the readers.

Signed-off-by: Daichi Mukai <daichi-mukai@cybozu.co.jp>