Fix concurrent map writes bug of K8s metrics API client
The pvc-autoresizer can use Prometheus or the K8s metrics API as sources
of metrics to determine whether a PVC needs to be resized. There is a
concurrent map access bug in the K8s metrics API client implementation,
which causes the pvc-autoresizer to crash. This commit fixes this issue.

Signed-off-by: Daichi Mukai <daichi-mukai@cybozu.co.jp>