Reduce image build time
We initially used `docker buildx build` to support multiple CPU
architectures[1]. However, we forgot to remove the make image command
that runs `docker build`, which was redundant, so we decided to
eliminate it. Another inefficiency arose from setting the branch tag.
The image is conceptually the same as the fully versioned one, but we
were running another `docker buildx build`, resulting in the creation of
an additional image. To address this, we now use `docker buildx
imagetools create` to set the branch tag on the existing image.

[1]: 99b821c (Adds multi-platform support (and ARM64 compat))

Signed-off-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>